adapt mon.pp and osd.pp to changed keyhandling
Don't create keys on your own in mon.pp, use ceph::key for this.

Don't generate client.admin key in ceph::mon and distribute it later
via facter to the other nodes. Simplify the handling of the secret for
client.admin: generate it like the mon secret before in the site.pp
and pass the secret over to the nodes which may need the key to generate
a file containing the key.

With this change the client.admin key is injected in the cluster
in ceph::mon. Don't care which of the MONs inject, since they all
use the same secret. This way it would reduce the needed puppet runs
from at least 3 runs to simply one per MON node. It also make administration
much easier since the client.amin secret is known to the admin from
the site.pp and can easily be used to create new keys on other machines.

Due to this change the 'client_admin_secret' secret is mandatory for
ceph::mon and ceph::osd.

Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>